Patch#: 590 Submitter: agilejoe
Applied with modifications.  Adds a RescueAttribute for rendering a specified view when a specified exception is thrown when running an action.  Conventional location for Rescues changed to Views/Shared/Rescues/{view} Removed the .aspx extension to be nice to other ViewFactory implementations.  Added a sample to the ConventionController sample site.