Patch#: 640 Submitter: JeremyS
More enhancements to ConventionController and ControllerDescriptor. 1) Adds a HiddenAction attribute which can be used to prevent a public method on a controller from being made accessible as an Action (basically, the opposite of the ControllerAction attribute). 2) Adds a DefaultAction attribute which specifies an action to invoke if the action that was called does not exist. 3) Added a SelectedAction property to ConventionController. 4) Added example usage to the ConventionController Sample site.