Issue#: 785
Ported Brail ViewEngine from MonoRail project.  This is the first step in helping teams interested in migrating existing investments in MonoRail to ASP.Net MVC.  The code was intentionally modified as little as possible to help with the porting of future bug fixes.  Two major omissions from MonoRail are ViewComponents and JavaScript partials.  Extension methods are not yet supported.