BREAKING CHANGE FOR NVelocityViewFactory
The reason:
The way NVelocityViewFactory merges templates together, it wasn't possible to use #parse("partial.vm") and have the output rendered in the right spot.  This made componentization difficult (if not impossible).  From Jeremy Skinner....

    When you use a Layout with NVelocity, the view will no longer be
    rendered directly to the response stream (Response.Output), but will
    instead be buffered in an internal StringWriter. Once the view has
    finished rendering, this buffered output is then merged into the
    template for the masterpage which is then rendered to Response.Output.


The change:
We changed this behavior, so now the child template no longer gets buffered into it's own stringwriter.  This allows us to just merge the master page and all the #parse statements together in one go.  So $childContent no longer has the entire content of the child view, instead it contains the path to the child view template.  This means that you'll need to wrap $childContent with #parse() to get it to work.  Again, to fix your sites, replace $childContent with #parse( $childContent ).

For more details, check the entire thread here:
http://groups.google.com/group/mvccontrib-discuss/browse_thread/thread/fc84d69db708322