Patch #3300 - Luizbon Applied with significant modifications:
The patch changed the behaviour of SingleFormat so that it was used when the total pages was equal to 1. This is incorrect - the numbers rendered in the SingleFormat represent the *page numbers* (eg "Showing 1 of 3" means that we're on page 1 out of a total of 3 pages). The patched version would instead render this message when the TotalPages was equal to 1 (with a page size of 3) which does not make sense as there is only 1 page! PaginationSingleFormat should only apply when the *PageSize* is 1. Therefore, I have not included this change.

However, I did include the other change from the patch (only show the pagination links if there is more than 1 page).